issue #9 : check if the composer tool exists

+++ build.sh
@@ -1,28 +1,40 @@
 #!/bin/bash
 
-echo "update the version of progpilot in console/application.php file"
-
 composertool=/home/eric/dev/phar-composer.phar
 version="dev"
 date=`date "+%Y%m%d-%H%M%S"`
 newfile="progpilot_${version}${date}"
     
-cd ./projects/phar
-rm composer.lock
-rm -rf ./vendor/
-composer install
+echo "Progpilot builder"
+echo "Have you updated the version of progpilot in Console/Application.php file ? (optional)"
 
-rm -rf ./vendor/progpilot/
-mkdir ./vendor/progpilot/
-mkdir ./vendor/progpilot/package
-cp -R ../../package/* ./vendor/progpilot/package
+if [ -e ${composertool} ] 
+then
+    cd ./projects/phar
+    rm composer.lock
+    rm -rf ./vendor/
+    composer install
 
-rm -rf ../../builds/*
-    
-echo "generating phar"
+    rm -rf ./vendor/progpilot/
+    mkdir ./vendor/progpilot/
+    mkdir ./vendor/progpilot/package
+    cp -R ../../package/* ./vendor/progpilot/package
 
-php -d phar.readonly=off ${composertool} build .
-mv ./rogpilot.phar ../../builds/${newfile}.phar
+    rm -rf ../../builds/*
+        
+    echo "generating phar"
 
-rm composer.lock
-rm -rd ./vendor/
+    php -d phar.readonly=off ${composertool} build .
+    mv ./rogpilot.phar ../../builds/${newfile}.phar
+
+    rm composer.lock
+    rm -rd ./vendor/
+else
+
+echo ""
+echo "Error : ${composertool} doesn't exist"
+echo "download the latest release of phar composer tool : https://github.com/clue/phar-composer/releases"
+echo "and define the path of the tool in the composertool variable of this script"
+
+fi
+
